refector: 배송 담당자 생성 로직 도메인 변경 (delivery -> auth)
- 배송 담당자 생성시 필요한 시퀀스 생성 로직 추가
